# Conflicts:
#	back/server/src/main/kotlin/com/metamong/common/util/DateExtensions.kt

+++ back/server/src/main/kotlin/com/metamong/infra/persistence/apartment/projection/ApartmentTradeChartProjection.kt
@@ -1,8 +1,8 @@
 package com.metamong.infra.persistence.apartment.projection
 
 data class ApartmentTradeChartProjection(
-    val contractYear: Short,
-    val contractMonth: Short,
+    val contractYear: Int,
+    val contractMonth: Int,
     val avgPrice: Double,
     val maxPrice: Int,
     val minPrice: Int,
